numFrobEntities should be int and not float.
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@321 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ game/player.cpp
@@ -7,6 +7,9 @@
  * $Author$
  *
  * $Log$
+ * Revision 1.46  2006/01/27 14:07:16  sparhawk
+ * numFrobEntities should be int and not float.
+ *
  * Revision 1.45  2006/01/24 22:03:46  sparhawk
  * Stim/Response implementation preliminary
  *
@@ -6949,9 +6952,9 @@ void idPlayer::Think( void )
 	FrobBounds.ExpandSelf( g_Global.m_MaxFrobDistance );
 	idEntity *FrobRangeEnts[ MAX_GENTITIES ];
 
-	float numFrobEnt = gameLocal.clip.EntitiesTouchingBounds( FrobBounds, -1, FrobRangeEnts, MAX_GENTITIES );
+	int numFrobEnt = gameLocal.clip.EntitiesTouchingBounds( FrobBounds, -1, FrobRangeEnts, MAX_GENTITIES );
 
-	for( int i=0; i<numFrobEnt; i++ )
+	for( int i=0; i < numFrobEnt; i++ )
 	{
 		if( FrobRangeEnts[i] )
 		{
